% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login Function to Validate Password}
\usage{
login(input, output, session, log_out = NULL)
}
\arguments{
\item{input}{Session input}

\item{output}{Session output}

\item{session}{Session info}

\item{log_out}{Logout check}
}
\value{
Authorized Credentials
}
\description{
This function is the Password Validation and other checks.
}
\details{
Requires: shinyjs, sodium
}
\examples{

}
