% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginUI.R
\name{loginUI}
\alias{loginUI}
\title{Login UI to Render Sign In Page}
\usage{
loginUI(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Sign in Page
}
\description{
This function is the Login UI to Render Sign In Page.
}
\details{
Requires: shinyjs
}
\examples{

}
