% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout Function}
\usage{
logout(input, output, session, active)
}
\arguments{
\item{input}{Session input}

\item{output}{Session output}

\item{session}{Session info}

\item{active}{Active check}
}
\value{
Updated Credentials
}
\description{
This function process the Logout from the application.
}
\details{
Requires: shinyjs, sodium
}
\examples{

}
