% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoutUI.R
\name{logoutUI}
\alias{logoutUI}
\title{LogoutUI Function}
\usage{
logoutUI(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Log In Panel
}
\description{
This function processes the Log Out on button click.
}
\details{
Requires: shinyjs
}
\examples{

}
